dev_requirements.txt
Django==4.0.3
django-ical==1.8.3
django_debug_toolbar==3.2.4
pytest-django
black
Fabric<2.0